+++ src/app/create-order/page.tsx
@@ -32,9 +32,9 @@ import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 import { useRouter } from "next/navigation";
 import Cookies from "js-cookie";
 import { jwtDecode } from 'jwt-decode';
-import Layout from "../../components/Layout";
-import OrderService from "../../../services/OrderService";
-import MenuService from "../../../services/MenuService";
+import Layout from "@/src/components/Layout";
+import OrderService from "@/src/services/OrderService";
+import MenuService from "@/src/services/MenuService";
 
 interface User {
   username?: string;
