+++ src/app/profile/page.tsx
@@ -5,8 +5,8 @@ import Cookies from 'js-cookie';
 import { jwtDecode } from 'jwt-decode';
 import { Box, Typography, Paper, Stack, Divider } from '@mui/material';
 import Layout from '@/src/components/Layout';
-import MenuService from '@/services/MenuService';
-import OrderService from '@/services/OrderService';
+import MenuService from '@/src/services/MenuService';
+import OrderService from '@/src/services/OrderService';
 
 interface User {
   username?: string;
